Clean up the insightToPropsConverter code
Split core and convenience functions

JIRA: RAIL-4069
