+++ libs/sdk-ui-ext/src/internal/utils/embeddingCodeGenerator/insightToPropsConverter/index.ts
@@ -1,0 +1,10 @@
+// (C) 2022 GoodData Corporation
+export {
+    bucketConversion,
+    IBucketConversion,
+    getInsightToPropsConverter,
+    insightConversion,
+    IInsightConversion,
+    ConversionSpec,
+} from "./convertor";
+export * from "./convenience";
